# apps/api/src/review_api/static/openapi.yaml
openapi: 3.1.0
info:
  title: Review Platform API
  version: 1.0.2
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Design-first contract between the React web client and backend for immutable
    document versions, asynchronous review, immutable reports, per-finding
    dialogue, and human decisions. The deployment provides one trusted actor and
    one configured organization/workspace. All timestamps are UTC RFC 3339.
    Review reports never contain mutable dialogue or decisions.
servers:
  - url: /api
    description: Same-origin reverse proxy
security: []
...
paths:
...
  /v1/workspaces/{workspaceId}/review-runs/{runId}/findings/{findingId}/dialogue:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/RunId"
      - $ref: "#/components/parameters/FindingId"
    get:
      tags: [Finding dialogue]
      operationId: getFindingDialogue
      summary: Return ordered dialogue and server-computed ability to send a message
      responses:
        "200":
          description: Current finding dialogue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingDialogue"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/ReportUnavailable"
  /v1/workspaces/{workspaceId}/review-runs/{runId}/findings/{findingId}/dialogue/turns:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/RunId"
      - $ref: "#/components/parameters/FindingId"
    post:
      tags: [Finding dialogue]
      operationId: createFindingDialogueTurn
      summary: Add one actor message and enqueue exactly one assistant response
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDialogueTurn"
      responses:
        "202":
          description: Turn accepted or replayed; poll the returned dialogue
          headers:
            Location:
              schema: {type: string, format: uri-reference}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingDialogue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/workspaces/{workspaceId}/review-runs/{runId}/findings/{findingId}/dialogue/turns/{turnId}/retry:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/RunId"
      - $ref: "#/components/parameters/FindingId"
      - $ref: "#/components/parameters/TurnId"
    post:
      tags: [Finding dialogue]
      operationId: retryFindingDialogueTurn
      summary: Retry generation for a failed turn without adding a new message
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetryDialogueTurn"
      responses:
        "202":
          description: Retry accepted or replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingDialogue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/workspaces/{workspaceId}/review-runs/{runId}/findings/{findingId}/decision:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/RunId"
      - $ref: "#/components/parameters/FindingId"
    put:
      tags: [Review runs]
      operationId: putFindingDecision
      summary: Record or reset a human decision with optimistic concurrency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutFindingDecision"
      responses:
        "200":
          description: Current human decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumanDecision"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      description: >-
        Must equal the deployment-configured workspace ID. Any other value is an
        ordinary namespace miss and returns 404.
      schema: {type: string, format: uuid}
...
    RunId:
      name: runId
      in: path
      required: true
      schema: {type: string, format: uuid}
    FindingId:
      name: findingId
      in: path
      required: true
      schema: {type: string, format: uuid}
    TurnId:
      name: turnId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Unique within workspace and operation for at least 24 hours. Same key and
        body replays the original result; same key with a different body returns 409.
      schema: {type: string, minLength: 8, maxLength: 128}
  responses:
    BadRequest:
      description: Request does not satisfy the contract
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource does not exist in the configured workspace namespace
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: State, idempotency, active-turn, or revision conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ReportUnavailable:
      description: Run has no published report yet
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Actor:
      type: object
      additionalProperties: false
      description: Deployment-configured identity used only for attribution
      required: [id, display_name]
      properties:
        id: {type: string, format: uuid}
        display_name: {type: string, minLength: 1, maxLength: 255}
...
    SkillSnapshot:
      type: object
      additionalProperties: false
      required: [id, version, package_sha256]
      properties:
        id: {type: string, minLength: 1, maxLength: 128}
        version: {$ref: "#/components/schemas/SemVer"}
        package_sha256: {type: string, pattern: "^[a-f0-9]{64}$"}
...
    FindingDialoguePolicy:
      type: object
      additionalProperties: false
      required: [id, version, digest, max_member_turns]
      properties:
        id: {type: string, minLength: 1, maxLength: 128}
        version: {$ref: "#/components/schemas/SemVer"}
        digest: {type: string, pattern: "^[a-f0-9]{64}$"}
        max_member_turns:
          type: [integer, "null"]
          minimum: 1
          description: Null means no numeric limit; other state/access rules still apply
...
    EvidenceAnchor:
      type: object
      additionalProperties: false
      required: [source_id, document_id, source_name, fragment_id, quote, quote_start, quote_end, location]
      properties:
        source_id: {type: string, minLength: 1, maxLength: 128}
        document_id: {type: string, format: uuid}
        source_name: {type: string, minLength: 1, maxLength: 255}
        fragment_id: {type: string, minLength: 1, maxLength: 160}
        quote: {type: string, minLength: 1, maxLength: 4000}
        quote_start: {type: integer, minimum: 0}
        quote_end: {type: integer, minimum: 1}
        location:
          $ref: "#/components/schemas/SourceLocation"
    SourceLocation:
      oneOf:
        - $ref: "#/components/schemas/PdfLocation"
        - $ref: "#/components/schemas/TextLocation"
      discriminator:
        propertyName: kind
        mapping:
          pdf: "#/components/schemas/PdfLocation"
          text: "#/components/schemas/TextLocation"
    PdfLocation:
      type: object
      additionalProperties: false
      required: [kind, page, rects]
      properties:
        kind: {const: pdf}
        page: {type: integer, minimum: 1}
        rects:
          type: array
          minItems: 1
          items:
            type: array
            prefixItems:
              - {type: number, minimum: 0, maximum: 1}
              - {type: number, minimum: 0, maximum: 1}
              - {type: number, minimum: 0, maximum: 1}
              - {type: number, minimum: 0, maximum: 1}
            items: false
        table: {type: [integer, "null"], minimum: 1}
        row: {type: [integer, "null"], minimum: 1}
    TextLocation:
      type: object
      additionalProperties: false
      required: [kind, line_start, line_end, char_start, char_end]
      properties:
        kind: {const: text}
        line_start: {type: integer, minimum: 1}
        line_end: {type: integer, minimum: 1}
        char_start: {type: integer, minimum: 0}
        char_end: {type: integer, minimum: 1}
...
    HumanDecision:
      type: object
      additionalProperties: false
      required: [status, revision, actor, reason, resolution, decided_at]
      properties:
        status: {enum: [unreviewed, confirmed, rejected, needs_context]}
        revision: {type: integer, minimum: 0}
        actor:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/Actor"
        reason: {type: [string, "null"], maxLength: 4000}
        resolution: {type: [string, "null"], maxLength: 8000}
        decided_at: {type: [string, "null"], format: date-time}
      allOf:
        - if:
            properties: {status: {const: unreviewed}}
            required: [status]
          then:
            properties:
              actor: {type: "null"}
              reason: {type: "null"}
              resolution: {type: "null"}
              decided_at: {type: "null"}
          else:
            properties:
              actor: {$ref: "#/components/schemas/Actor"}
              reason: {type: string, minLength: 1, maxLength: 4000}
              decided_at: {type: string, format: date-time}
    PutFindingDecision:
      type: object
      additionalProperties: false
      required: [status, reason, resolution, expected_revision]
      properties:
        status: {enum: [unreviewed, confirmed, rejected, needs_context]}
        reason: {type: [string, "null"], maxLength: 4000}
        resolution: {type: [string, "null"], maxLength: 8000}
        expected_revision: {type: integer, minimum: 0}
      allOf:
        - if:
            properties: {status: {const: unreviewed}}
            required: [status]
          then:
            properties:
              reason: {type: "null"}
              resolution: {type: "null"}
          else:
            properties:
              reason: {type: string, minLength: 1, maxLength: 4000}
...
    FindingDialogue:
      type: object
      additionalProperties: false
      required: [id, run_id, finding_id, revision, state, turn_count, can_send_message, blocked_reason, policy, turns]
      properties:
        id: {type: string, format: uuid}
        run_id: {type: string, format: uuid}
        finding_id: {type: string, format: uuid}
        revision: {type: integer, minimum: 0}
        state: {enum: [open, generating, closed]}
        turn_count: {type: integer, minimum: 0}
        can_send_message: {type: boolean}
        blocked_reason:
          type: [string, "null"]
          enum: [generation_in_progress, turn_limit_reached, human_decision_recorded, dialogue_not_supported, model_unavailable, null]
        policy:
          $ref: "#/components/schemas/FindingDialoguePolicy"
        turns:
          type: array
          items:
            $ref: "#/components/schemas/DialogueTurn"
    DialogueTurn:
      type: object
      additionalProperties: false
      required: [id, ordinal, state, actor, member_message, created_at, assistant_response, error, finished_at]
      properties:
        id: {type: string, format: uuid}
        ordinal: {type: integer, minimum: 1}
        state: {enum: [queued, generating, completed, failed]}
        actor:
          $ref: "#/components/schemas/Actor"
        member_message: {type: string, minLength: 1, maxLength: 8000}
        created_at: {type: string, format: date-time}
        assistant_response:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/AssistantResponse"
        error:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/DialogueError"
        finished_at: {type: [string, "null"], format: date-time}
    AssistantResponse:
      type: object
      additionalProperties: false
      required: [action, content, proposed_resolution, anchors, provenance]
      properties:
        action: {enum: [clarify, propose_resolution, escalate]}
        content: {type: string, minLength: 1, maxLength: 8000}
        proposed_resolution:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/ProposedResolution"
        anchors:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceAnchor"
        provenance:
          $ref: "#/components/schemas/AssistantProvenance"
    ProposedResolution:
      type: object
      additionalProperties: false
      required: [text, rationale]
      properties:
        text: {type: string, minLength: 1, maxLength: 8000}
        rationale: {type: string, minLength: 1, maxLength: 4000}
    AssistantProvenance:
      type: object
      additionalProperties: false
      required: [skill, model]
      properties:
        skill:
          $ref: "#/components/schemas/SkillSnapshot"
        model:
          $ref: "#/components/schemas/ModelExecution"
    DialogueError:
      type: object
      additionalProperties: false
      required: [code, message, retryable]
      properties:
        code: {enum: [model_unavailable, context_limit, content_blocked, model_output_invalid, validation_failed, internal_error]}
        message: {type: string, minLength: 1, maxLength: 1000}
        retryable: {type: boolean}
    CreateDialogueTurn:
      type: object
      additionalProperties: false
      required: [message, expected_revision]
      properties:
        message: {type: string, minLength: 1, maxLength: 8000}
        expected_revision: {type: integer, minimum: 0}
    RetryDialogueTurn:
      type: object
      additionalProperties: false
      required: [expected_revision]
      properties:
        expected_revision: {type: integer, minimum: 0}
...
    ModelExecution:
      type: object
      additionalProperties: false
      required: [provider, model, model_version, safe_parameters, usage]
      properties:
        provider: {type: string, minLength: 1, maxLength: 128}
        model: {type: string, minLength: 1, maxLength: 255}
        model_version: {type: string, minLength: 1, maxLength: 255}
        safe_parameters:
          type: object
          additionalProperties:
            type: [string, number, integer, boolean, "null"]
        usage:
          type: object
          additionalProperties: false
          required: [input_tokens, output_tokens]
          properties:
            input_tokens: {type: [integer, "null"], minimum: 0}
            output_tokens: {type: [integer, "null"], minimum: 0}
...
    SemVer:
      type: string
      pattern: "^[0-9]+\\.[0-9]+\\.[0-9]+(?:[-+][0-9A-Za-z.-]+)?$"
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string, minLength: 1, maxLength: 255}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string, maxLength: 4000}
        instance: {type: string, format: uri-reference}
        code: {type: string, pattern: "^[a-z][a-z0-9_]*$"}
        request_id: {type: string, minLength: 1, maxLength: 128}
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, message]
            properties:
              field: {type: string, minLength: 1, maxLength: 255}
              message: {type: string, minLength: 1, maxLength: 1000}